+++ lib/src/base32.dart
@@ -76,5 +76,5 @@ Uint8List fromBase32(
 }) {
   codec ??= _codecFromParameters(padding: padding);
   var out = codec.decoder.convert(input.codeUnits);
-  return Uint8List.fromList(out.toList());
+  return Uint8List.fromList(out as List<int>);
 }