+++ lib/src/base64.dart
@@ -73,5 +73,5 @@ Uint8List fromBase64(
 }) {
   codec ??= _codecFromParameters(padding: padding);
   var out = codec.decoder.convert(input.codeUnits);
-  return Uint8List.fromList(out.toList());
+  return Uint8List.fromList(out as List<int>);
 }